+++ src/anime_parsers_ru/__init__.py
@@ -1,4 +1,5 @@
 from .parser_kodik import KodikParser
+from .parser_kodik_async import KodikParserAsync
 from .parser_aniboom import AniboomParser
 from .parser_jutsu import JutsuParser
 from .parser_shikimori import ShikimoriParser